+++ detekt-rules-naming/src/main/kotlin/io/gitlab/arturbosch/detekt/rules/naming/NamingProvider.kt
@@ -32,6 +32,10 @@ class NamingProvider : DefaultRuleSetProvider {
             FunctionNaming(config),
             EnumNaming(config),
             ObjectPropertyNaming(config),
+            FunctionMinLength(config),
+            FunctionMaxLength(config),
+            VariableMaxLength(config),
+            VariableMaxLength(config),
         )
     )
 }
