--- detekt-rules-naming/src/test/kotlin/io/gitlab/arturbosch/detekt/rules/naming/NamingRulesSpec.kt
@@ -1,19 +1,0 @@
-package io.gitlab.arturbosch.detekt.rules.naming
-
-import io.gitlab.arturbosch.detekt.test.assertThat
-import io.gitlab.arturbosch.detekt.test.compileAndLint
-import org.junit.jupiter.api.Test
-
-class NamingRulesSpec {
-
-    @Test
-    fun `should not detect any`() {
-        val code = """
-            data class D(val i: Int, val j: Int)
-            fun doStuff() {
-                val (_, HOLY_GRAIL) = D(5, 4)
-            }
-        """
-        assertThat(NamingRules().compileAndLint(code)).isEmpty()
-    }
-}
